requirements.txt
Flask
flask-cors
google-cloud-bigquery
google-cloud-aiplatform
python-dotenv
vertexai
gunicorn